add appAction for opening tree context menu
